divided functions into 3 files
