+++ src/helpers.py
@@ -1,4 +1,20 @@
-# import the dependencies
+'''
+Helper functions for processing and analyzing single-cell (scRNA-seq) and single-nucleus (snRNA-seq) data.
+
+Includes utilities for:
+- Loading and preprocessing raw and compressed datasets
+- Cleaning and harmonizing metadata and cell type annotations
+- Generating pseudobulks and reference matrices for deconvolution
+- Quality control checks across data sources
+- Model training (e.g., scVI) and evaluation metrics
+
+Supports multiple datasets (PNB, OVC, MBC, Real_ADP) and formats outputs for tools like CIBERSORT and BayesPrism.
+
+This script contains helper functions needed throughout the repository. 
+It should be in /src/ folder and imported as "from src.helpers import the_function"
+'''
+
+# Import the dependencies
 import sys
 from sklearn.decomposition import PCA
 from sklearn.preprocessing import StandardScaler
@@ -20,13 +36,10 @@ from sklearn.utils import resample
 from scipy.spatial.distance import squareform
 from scipy.cluster.hierarchy import linkage, dendrogram
 import matplotlib.pyplot as plt
+import json
 import seaborn as sns
-import pickle
 import scanpy as sc
 from scipy.sparse import issparse
-from pydeseq2.default_inference import DefaultInference
-from pydeseq2.dds import DeseqDataSet
-from pydeseq2.ds import DeseqStats
 import numpy as np
 import pandas as pd
 from numpy.linalg import norm
@@ -43,390 +56,17 @@ sys.path.insert(1, '../../')
 sys.path.insert(1, '../')
 sys.path.insert(1, '../../../')
 
-def transform_heldout_sn_to_mean_sc_VAE(
-    model: scvi.model.SCVI,
-    sc_adata: sc.AnnData,
-    sn_adata: sc.AnnData,
-    sn_heldout_adata: sc.AnnData,
-    heldout_label: str,
-    k_neighbors: int = 10
-):
+def compute_metrics(y_true, y_pred):
     """
-    Shift single-nucleus (SN) held-out cells into a single-cell (SC)-like expression space using a local neighbor-based
-    transformation in the scVI latent space.
-
-    This function aims to transform the held-out SN cells so that their latent representations resemble those of SC cells.
-    It does so by:
-    
-    1. **Gene Alignment**: Subsets each AnnData object to the set of genes common to sc_adata, sn_adata, and sn_heldout_adata.
-    2. **Latent Representation**: Uses the provided scVI model to obtain latent embeddings (z_sc, z_sn, and z_ho) for the SC data,
-       the non-held-out SN data, and the held-out SN data, respectively.
-    3. **Local Shift Computation**:
-       - Computes a shift vector for each non-held-out SN cell as (mean latent SC) – (latent SN).
-         The mean latent SC is the centroid (mean) of the SC embeddings.
-       - For each held-out SN cell, finds its k nearest neighbors (in latent space) among the non-held-out SN embeddings
-         and averages their shift vectors. This local shift is then added to the held-out cell’s latent representation,
-         yielding a “SC-like” latent position.
-    4. **Decode to Expression**: After updating the held-out SN cells’ latent representations, it decodes these shifted
-       embeddings back to gene expression space with a library size scaled to the median of the SC data.
-    5. **Output**: Returns a DataFrame of the transformed expression matrix (rows = held-out SN cells, columns = genes).
-
-    Parameters
-    ----------
-    model : scvi.model.SCVI
-        A trained scVI model (non-conditional). It should be fit on SC plus SN minus the held-out cell type.
-    sc_adata : sc.AnnData
-        Single-cell AnnData, excluding the held-out cell type. Used to define the SC latent centroid and library size.
-    sn_adata : sc.AnnData
-        Single-nucleus AnnData, excluding the held-out cell type. Used to build the neighbor graph and compute local shifts.
-    sn_heldout_adata : sc.AnnData
-        The held-out single-nucleus cells, which will be transformed into a SC-like latent space.
-    heldout_label : str
-        The label/name of the held-out cell type. Not currently used inside the function,
-        but kept for record or future reference in 'sn_heldout_adata.obs'.
-    k_neighbors : int, default=10
-        The number of nearest neighbors to use for local shift averaging.
-
-    Returns
-    -------
-    pd.DataFrame
-        A DataFrame containing the shifted and decoded gene expression for the held-out SN cells. Each row corresponds
-        to one held-out cell, and columns are the common genes shared by SC, SN, and the held-out subset.
-
-    Notes
-    -----
-    - This procedure is designed to emulate a local “shift” from SN to SC in the scVI latent space, preserving
-      local structure among the SN cells while anchoring to the global SC centroid.
-    - The scVI model should not have been trained on the held-out cell type (i.e., that cell type’s SN cells
-      were excluded from model training) to avoid data leakage and simulate a real missing-cell scenario.
+    Compute RMSE and Pearson correlation (or NaN if fewer than 2 points).
+    Returns (rmse, pearson).
     """
-
-    # 1) Align genes
-    common_genes = set(sc_adata.var_names).intersection(sn_adata.var_names).intersection(sn_heldout_adata.var_names)
-    sc_adata = sc_adata[:, list(common_genes)].copy()
-    sn_adata = sn_adata[:, list(common_genes)].copy()
-    sn_heldout_adata = sn_heldout_adata[:, list(common_genes)].copy()
-
-    sc_celltype_labels = sc_adata.obs["cell_types"].astype(str)
-    sn_celltype_labels = sn_adata.obs["cell_types"].astype(str)
-
-    # Identify overlapping cell types (excluding the held-out type)
-    overlapping_types = np.intersect1d(sc_celltype_labels.unique(), sn_celltype_labels.unique())
-    overlapping_types = overlapping_types[overlapping_types != heldout_label]
-    
-    sc_data_aligned = sc_adata[sc_adata.obs.cell_types.isin(overlapping_types)].copy().to_df()
-    sn_data_aligned = sn_adata[sn_adata.obs.cell_types.isin(overlapping_types)].copy().to_df()
-
-    # combined SC + SN(overlapping), excluding the held-out SN
-    keep_sn_mask = (sn_celltype_labels != heldout_label)
-
-    # 2) Get latent representations
-    z_sc = model.get_latent_representation(sc_adata)
-    z_sn = model.get_latent_representation(sn_adata)
-
-    mean_sc_latent = z_sc.mean(axis=0)
-    sn_shifts = mean_sc_latent - z_sn  # shape: (#sn_cells, latent_dim)
-
-    z_ho = model.get_latent_representation(sn_heldout_adata)
-
-    # 3) Build kNN on sn latent space
-    knn = NearestNeighbors(n_neighbors=k_neighbors).fit(z_sn)
-    z_ho_shifted = np.zeros_like(z_ho)
-
-    for i in range(z_ho.shape[0]):
-        _, neigh_idx = knn.kneighbors(z_ho[i, :].reshape(1, -1))
-        local_shift = sn_shifts[neigh_idx[0]].mean(axis=0)
-        z_ho_shifted[i, :] = z_ho[i, :] + local_shift
-
-    # 4) Put z_ho_shifted in sn_heldout_adata.obsm
-    sn_heldout_adata.obsm["_scvi_latent"] = z_ho_shifted
-
-    # 5) Determine library size from sc_adata
-    sc_totals = np.array(sc_adata.X.sum(axis=1)).flatten()
-    median_sc_lib = np.median(sc_totals)
-
-    transformed_expr = get_normalized_expression_from_latent(
-        model,
-        z_ho_shifted,
-        library_size=median_sc_lib,
-        gene_list=list(common_genes),
-        return_numpy=True
-    )
-    
-    # 7) Return as a DataFrame with the same shape as the # of cells in sn_heldout_adata
-    #    and columns = common_genes
-    df_out = pd.DataFrame(
-        transformed_expr,
-        index=sn_heldout_adata.obs_names,
-        columns=list(common_genes)
-    )
-    return df_out
-
-def get_normalized_expression_from_latent(
-    model: scvi.model.SCVI,
-    latent: np.ndarray | torch.Tensor,
-    library_size: float | None = 1.0,
-    gene_list: list[str] | None = None,
-    n_samples: int = 1,
-    return_mean: bool = True,
-    return_numpy: bool = True,
-):
-    """
-    Decode expression from a user-provided latent representation using scvi-tools 1.2.x,
-    which requires calling model.module.decoder(x, dispersion, library).
-
-    Parameters
-    ----------
-    model : scvi.model.SCVI
-        A trained scVI model. Typically configured with `dispersion="gene-batch"` or similar.
-    latent : np.ndarray or torch.Tensor
-        The (n_cells, n_latent) latent representation to decode. This could be your
-        custom-shifted latent for hold-out cells.
-    library_size : float or None, default=1.0
-        If not None, each cell's px_scale is multiplied by this value.
-        By default=1.0, all cells have library=1.0. If you prefer no uniform scaling,
-        pass None and then do any scaling afterwards, or pass a different float to enforce
-        that uniform sum. 
-    gene_list : list of str, optional
-        If provided, only return these genes. Must exist in `model.adata.var_names`.
-    n_samples : int, default=1
-        Number of times to sample from the decoder. If > 1 and `return_mean=True`,
-        the output is the average expression across samples.
-    return_mean : bool, default=True
-        If True and `n_samples>1`, returns the mean. If False and n_samples>1,
-        returns a 3D (n_samples, n_cells, n_genes) array or tensor.
-    return_numpy : bool, default=True
-        If True, returns a NumPy array; otherwise returns a pandas DataFrame
-        (with shape `(n_cells, n_genes)` or `(n_samples, n_cells, n_genes)`).
-
-    Returns
-    -------
-    np.ndarray or pd.DataFrame
-        Decoded expression from the latent codes. If `return_mean=True`, shape is (n_cells, n_genes).
-        If `return_mean=False and n_samples>1`, shape is (n_samples, n_cells, n_genes).
-
-    Notes
-    -----
-    - This function bypasses the encoder stage. It only decodes from the provided `latent` matrix.
-    - scVI's `dispersion` mode is inferred from the model itself, e.g. `model.module.dispersion`.
-    - The shape of `latent` must be `(n_cells, n_latent)`.
-    """
-    device = next(model.module.decoder.parameters()).device
-    dispersion_mode = model.module.dispersion
-    # 1) Ensure 'latent' is on the correct device, shape=(n_cells, n_latent)
-    if not torch.is_tensor(latent):
-        latent_tensor = torch.tensor(latent, dtype=torch.float32, device=device)
+    rmse = root_mean_squared_error(y_true, y_pred)
+    if len(y_true) < 2:
+        pearson = np.nan
     else:
-        latent_tensor = latent.to(device)
-
-    # 2) Build library tensor of shape (n_cells, 1). 
-    #    If library_size=1.0 => library=1 for all cells. 
-    #    If None => you could revert to library=1 anyway, or pass 0 => but None means "no uniform scaling."
-    n_cells = latent_tensor.shape[0]
-    if library_size is not None:
-        library_tensor = torch.full((n_cells, 1), library_size, dtype=torch.float32, device=device)
-    else:
-        # default to all ones if you want "no uniform scaling" 
-        library_tensor = torch.ones((n_cells, 1), dtype=torch.float32, device=device)
-
-    # 3) Fetch the model's dispersion string. e.g. "gene-batch", "gene", etc.
-    dispersion_mode = model.module.dispersion
-    print("dispersion mode: ", dispersion_mode)
-
-    # 4) We'll accumulate multiple samples of px_scale if n_samples>1
-    samples_list = []
-    model.module.decoder.eval()  # turn off dropout, etc.
-    with torch.no_grad():
-        for _ in range(n_samples):
-            # scvi-tools 1.2.x expects DecoderSCVI.forward(x, dispersion, library)
-            # in that exact positional order:
-            if len(model.module.decoder.px_decoder.n_cat_list) > 0:
-                # For each expected categorical input, create a dummy tensor.
-                # Assuming each categorical covariate is of size (n_cells, 1)
-                dummy_cats = [torch.zeros((latent_tensor.size(0), 1), dtype=torch.long, device=device)
-                            for _ in model.module.decoder.px_decoder.n_cat_list]
-                decoder_out = model.module.decoder(
-                    dispersion_mode, latent_tensor, library_tensor, *dummy_cats
-                )
-            else:
-                decoder_out = model.module.decoder(
-                    dispersion_mode, latent_tensor, library_tensor)
-
-            # decoder_out is typically a dict with "px_scale", "px_r", etc.
-            px_scale = decoder_out[0]
-            samples_list.append(px_scale.cpu())
-
-    # 5) If multiple samples, handle them
-    if n_samples > 1:
-        stacked = torch.stack(samples_list, dim=0)  # shape: (n_samples, n_cells, n_genes)
-        if return_mean:
-            expr_tensor = stacked.mean(dim=0)        # shape: (n_cells, n_genes)
-        else:
-            expr_tensor = stacked                   # shape: (n_samples, n_cells, n_genes)
-    else:
-        expr_tensor = samples_list[0]              # shape: (n_cells, n_genes)
-
-    # 6) Subset genes if needed
-    var_names = model.adata.var_names
-    if gene_list is not None:
-        gene_indices = [i for i, g in enumerate(var_names) if g in gene_list]
-        if expr_tensor.ndim == 2:
-            expr_tensor = expr_tensor[:, gene_indices]
-        elif expr_tensor.ndim == 3:
-            expr_tensor = expr_tensor[..., gene_indices]
-        var_names = var_names[gene_indices]
-
-    # 7) Return as numpy or DataFrame
-    expr_np = expr_tensor.numpy()
-    if return_numpy:
-        return expr_np
-    else:
-        # If shape is (n_cells, n_genes) => DataFrame with gene columns
-        # If shape is (n_samples, n_cells, n_genes), we'd have to decide indexing. 
-        # Usually you only do DataFrame for 2D. 
-        if expr_np.ndim == 2:
-            return pd.DataFrame(expr_np, columns=var_names)
-        else:
-            # We'll just return a 3D np.array if shape is 3D
-            # or create a multi-index. It's simpler to keep it as np.
-            raise ValueError(
-                "DataFrame output for 3D arrays is ambiguous. "
-                "Set return_mean=True or return_numpy=True."
-            )
-
-def transform_heldout_sn_to_mean_sc_local(
-    sc_data: pd.DataFrame,
-    sn_data: pd.DataFrame,
-    sn_heldout_data: pd.DataFrame,
-    sc_celltype_labels: pd.Series,
-    sn_celltype_labels: pd.Series,
-    heldout_label: str,
-    variance_threshold: float = 0.90,
-    k_neighbors: int = 10
-):
-    """
-    Local PCA-based transformation for the 'held-out' SN cell type.
-
-    Unlike the original method that used a single “composite shift,”
-    this version does a local kNN in PCA space for each held-out cell.
-
-    Steps:
-    1) Align columns (genes) across SC, SN, and SN-heldout.
-    2) Fit PCA on combined SC + SN (excluding the held-out label).
-    3) For each overlapping SN cell, SHIFT_i = (mean(SC PCA) - sn_pcs[i]).
-    4) For each held-out SN cell:
-       - transform to PCA coords
-       - find k nearest neighbors among sn_pcs
-       - average SHIFT among those neighbors
-       - apply SHIFT -> invert transform -> exp -> clamp -> library-size scale
-    5) Return the resulting DataFrame with “SC-like” expression.
-
-    Parameters
-    ----------
-    sc_data : pd.DataFrame
-        Single-cell data, rows = cells, columns = genes (raw counts).
-    sn_data : pd.DataFrame
-        Single-nucleus data, rows = cells, columns = genes (raw counts),
-        for all SN cell types EXCEPT the held-out, plus the overlapping types.
-    sn_heldout_data : pd.DataFrame
-        Single-nucleus data for the held-out cell type, rows = cells, columns = genes (raw counts).
-    sc_celltype_labels : pd.Series
-        Cell-type labels for sc_data’s rows.
-    sn_celltype_labels : pd.Series
-        Cell-type labels for sn_data’s rows.
-    heldout_label : str
-        The SN cell type that is missing from SC but present in SN (sn_heldout_data).
-    variance_threshold : float
-        Fraction of variance to retain in PCA (e.g. 0.90).
-    k_neighbors : int
-        Number of neighbors for local averaging of SHIFT in PCA space.
-
-    Returns
-    -------
-    pd.DataFrame
-        The transformed SN-heldout data (rows = cells, columns = genes),
-        with expression scaled similarly to SC.
-    """
-    
-    # 1) Align columns
-    common_genes = sc_data.columns.intersection(sn_data.columns).intersection(sn_heldout_data.columns)
-    sc_data_aligned = sc_data[common_genes]
-    sn_data_aligned = sn_data[common_genes]
-    sn_heldout_aligned = sn_heldout_data[common_genes]
-
-    # 2) Identify overlapping cell types (excluding the held-out type)
-    overlapping_types = np.intersect1d(sc_celltype_labels.unique(), sn_celltype_labels.unique())
-    overlapping_types = overlapping_types[overlapping_types != heldout_label]
-
-    # 3) Fit PCA on combined SC + SN(overlapping), excluding the held-out SN
-    keep_sn_mask = (sn_celltype_labels != heldout_label)
-    combined_df = pd.concat([sc_data_aligned, sn_data_aligned.loc[keep_sn_mask]], axis=0)
-
-    # log1p + scale
-    combined_log = np.log1p(combined_df)
-    scaler = StandardScaler()
-    combined_scaled = scaler.fit_transform(combined_log)
-
-    # find # of comps for variance_threshold
-    # do PCA with chosen comps
-    pca = PCA(n_components=variance_threshold)
-    pca.fit(combined_scaled)
-
-    # transform SC
-    sc_log = np.log1p(sc_data_aligned)
-    sc_scaled = scaler.transform(sc_log)
-    sc_pcs = pca.transform(sc_scaled)
-
-    # transform SN (overlapping)
-    sn_overlap = sn_data_aligned.loc[keep_sn_mask]
-    sn_log = np.log1p(sn_overlap)
-    sn_scaled = scaler.transform(sn_log)
-    sn_pcs = pca.transform(sn_scaled)
-
-    # 4) SHIFT vectors for each overlapping SN cell
-    #    SHIFT_i = (mean(SC in PCA) - sn_pcs[i]).
-    mean_sc_pca = sc_pcs.mean(axis=0)
-    sn_shifts = mean_sc_pca - sn_pcs  # shape = (n_SN, n_comps)
-
-    # 5) local transform for held-out SN cells
-    #    (a) transform to PCA, (b) find KNN among sn_pcs, (c) average SHIFT, (d) apply SHIFT
-    ho_log = np.log1p(sn_heldout_aligned)
-    ho_scaled = scaler.transform(ho_log)
-    ho_pcs = pca.transform(ho_scaled)
-
-    # build knn on sn_pcs
-    knn = NearestNeighbors(n_neighbors=k_neighbors).fit(sn_pcs)
-    ho_pcs_shifted = np.zeros_like(ho_pcs)
-
-    for i in range(ho_pcs.shape[0]):
-        this_cell = ho_pcs[i, :].reshape(1, -1)
-        # find neighbors
-        _, neigh_idx = knn.kneighbors(this_cell)
-        # average SHIFT
-        local_shift = sn_shifts[neigh_idx[0], :].mean(axis=0)
-        ho_pcs_shifted[i, :] = ho_pcs[i, :] + local_shift
-
-    # invert transform: pca -> scale -> expm1
-    ho_log_transformed = pca.inverse_transform(ho_pcs_shifted)
-    ho_unscaled = scaler.inverse_transform(ho_log_transformed)
-    ho_counts = np.expm1(ho_unscaled)
-    ho_counts[ho_counts < 0] = 0
-
-    # 6) Scale each cell’s library to match median SC library size
-    sc_totals = sc_data_aligned.sum(axis=1)  # sum per cell
-    median_sc_lib = np.median(sc_totals)
-    for i in range(ho_counts.shape[0]):
-        row_sum = ho_counts[i, :].sum()
-        if row_sum > 0:
-            ho_counts[i, :] *= (median_sc_lib / row_sum)
-
-    # 7) Build final DataFrame
-    transformed_df = pd.DataFrame(
-        ho_counts, 
-        index=sn_heldout_aligned.index, 
-        columns=common_genes
-    )
-    return transformed_df
+        pearson, _ = pearsonr(y_true, y_pred)
+    return rmse, pearson
 
 def remove_unassigned_cells(adata, dataset_name):
     """
@@ -458,119 +98,6 @@ def remove_unassigned_cells(adata, dataset_name):
     print(f"{dataset_name}: Removed {removed_count} unassigned/missing cells. Remaining: {adata.n_obs}")
 
     return adata
-
-def transform_heldout_sn_to_mean_sc(
-    sc_data: pd.DataFrame,
-    sn_data: pd.DataFrame,
-    sn_heldout_data: pd.DataFrame,
-    sc_celltype_labels: pd.Series,
-    sn_celltype_labels: pd.Series,
-    heldout_label: str,
-    variance_threshold: float = 0.90):
-    """
-    Transforms the 'held-out' SN cell type (sn_heldout_data) to match the average
-    PCA profile shift across overlapping SC vs SN cell types, then scales
-    to match the median SC library size.
-
-    Parameters
-    ----------
-    sc_data : pd.DataFrame
-        Single-cell data, rows = cells, columns = genes (raw counts).
-    sn_data : pd.DataFrame
-        Single-nucleus data, rows = cells, columns = genes (raw counts).
-    sn_heldout_data : pd.DataFrame
-        Single-nucleus data for the held-out cell type, rows = cells, columns = genes (raw counts).
-    sc_celltype_labels : pd.Series
-        Cell-type labels for the rows of sc_data (index-aligned).
-    sn_celltype_labels : pd.Series
-        Cell-type labels for the rows of sn_data (index-aligned).
-    heldout_label : str
-        The label (cell type) that is missing in SC but present in SN (for sn_heldout_data).
-    variance_threshold : float
-        Fraction of variance to retain in PCA (e.g., 0.90).
-
-    Returns
-    -------
-    pd.DataFrame
-        The transformed SN-heldout data (rows = cells, columns = genes),
-        with matched gene columns and library sizes akin to typical SC.
-    """
-
-    # 1) Align columns (genes) across SC, SN, and SN-heldout
-    ## to ensure that only genes present in all three DataFrames are used. This step prevents misalignment in subsequent transforms.
-    common_genes = sc_data.columns.intersection(sn_data.columns).intersection(sn_heldout_data.columns)
-    sc_data_aligned = sc_data[common_genes]
-    sn_data_aligned = sn_data[common_genes]
-    sn_heldout_data_aligned = sn_heldout_data[common_genes]
-
-    # 2) Identify overlapping cell types (excluding the held-out type)
-    # We gather the cell types shared by both SC and SN.
-    overlapping_types = np.intersect1d(sc_celltype_labels.unique(), sn_celltype_labels.unique())
-    # We exclude the heldout type so it won’t affect the PCA transformation.
-    overlapping_types = overlapping_types[overlapping_types != heldout_label]
-
-    # 3) Fit PCA on combined SC+SN, excluding the held-out SN
-    ## We remove the heldout single‐nucleus cells from the PCA training set (i.e., “SN except heldout”), 
-    # so PCA is only learning from the types SC and SN have in common
-    keep_sn_mask = (sn_celltype_labels != heldout_label)
-    combined_df = pd.concat([sc_data_aligned, sn_data_aligned.loc[keep_sn_mask]], axis=0)
-
-    # Apply log1p and standard scaling
-    combined_log = np.log1p(combined_df)
-    scaler = StandardScaler()
-    combined_scaled = scaler.fit_transform(combined_log)
-
-    # Fit 
-    # PCA with the chosen variance threshold. 
-    # ensuring we only keep enough components to reach the variance_threshold
-    pca = PCA(n_components=variance_threshold)
-    pca.fit(combined_scaled)
-
-    # 4) Compute SHIFT VECTORS for each overlapping cell type
-    shift_vectors = []
-    for ct in overlapping_types:
-        sc_cells_ct = sc_data_aligned.loc[sc_celltype_labels == ct]
-        sn_cells_ct = sn_data_aligned.loc[sn_celltype_labels == ct]
-
-        if not sc_cells_ct.empty and not sn_cells_ct.empty:
-            sc_scaled = scaler.transform(np.log1p(sc_cells_ct))
-            sn_scaled = scaler.transform(np.log1p(sn_cells_ct))
-            sc_pcs = pca.transform(sc_scaled)
-            sn_pcs = pca.transform(sn_scaled)
-            shift_vectors.append(np.mean(sc_pcs, axis=0) - np.mean(sn_pcs, axis=0))
-
-    # 5) Compute the composite shift vector
-    if len(shift_vectors) == 0:
-        print("No overlapping cell types found! Using global SC-SN shift.")
-        all_sc_scaled = scaler.transform(np.log1p(sc_data_aligned))
-        all_sn_scaled = scaler.transform(np.log1p(sn_data_aligned.loc[keep_sn_mask]))
-        composite_shift = np.mean(pca.transform(all_sc_scaled), axis=0) - np.mean(pca.transform(all_sn_scaled), axis=0)
-    else:
-        composite_shift = np.mean(shift_vectors, axis=0)
-
-    # 6) Transform the HELD-OUT SN cells with that composite shift
-    ho_scaled = scaler.transform(np.log1p(sn_heldout_data_aligned))
-    ho_pcs = pca.transform(ho_scaled)
-    ho_pcs_shifted = ho_pcs + composite_shift
-    ho_log_transformed = pca.inverse_transform(ho_pcs_shifted)
-
-    # 7) Exponentiate (count scale) and clip negatives
-    # we revert the log1p transformation with expm1, then clamp negative values to zero.
-    ho_exp = np.expm1(ho_log_transformed)
-    ho_exp[ho_exp < 0] = 0  # Ensure no negatives
-
-    # 8) Scale each cell's library to match the median SC library size
-    # each cell in the transformed set is scaled to match the “typical” (median) SC library size.
-    median_sc_lib = np.median(sc_data_aligned.sum(axis=1))
-    for i in range(ho_exp.shape[0]):
-        row_sum = ho_exp[i, :].sum()
-        if row_sum > 0:
-            ho_exp[i, :] *= (median_sc_lib / row_sum)
-
-    # 9) Convert to DataFrame
-    transformed_df = pd.DataFrame(ho_exp, index=sn_heldout_data_aligned.index, columns=common_genes)
-
-    return transformed_df
 
 def filter_out_cell_markers(adata_sc, adata_sn, diff_genes, marker_threshold=50):
     """
@@ -1090,79 +617,6 @@ def qc_check_cell_types_match(original_anndata_path, sc_ref_path, sn_ref_path, p
 
     print("Cell types match across original AnnData, references, and pseudobulks.")
 
-def create_fixed_pseudobulk(adata, group_size=10):
-    """
-    Aggregates cells into pseudobulks of size `group_size` for each cell type.
-    If the cell count is not divisible by `group_size`, leftover cells are dropped.
-    
-    Returns:
-        AnnData where each row is a pseudobulk sample and columns are genes.
-        The new AnnData has:
-         - obs["cell_type"] to indicate which cell type the pseudobulk came from.
-         - obs["group_id"] to indicate the pseudobulk group number within that cell type.
-    """
-    import numpy as np
-    import pandas as pd
-    from anndata import AnnData
-    from scipy.sparse import issparse, csr_matrix
-    
-    all_pseudobulks = []
-    obs_list = []
-    var_names = adata.var_names
-    
-    # For each cell type, group the cells in sets of `group_size`
-    for cell_type in adata.obs["cell_types"].unique():
-        subset = adata[adata.obs["cell_types"] == cell_type]
-        n_cells = subset.shape[0]
-        
-        # Number of complete groups
-        n_groups = n_cells // group_size
-        if n_groups == 0:
-            continue  # skip if fewer than group_size cells for this type
-
-        # Only keep the cells that form complete groups
-        used_cells_count = n_groups * group_size
-        subset = subset[:used_cells_count, :]
-        
-        # Convert to dense or keep it sparse
-        mat = subset.X
-        if issparse(mat):
-            mat = mat.toarray()
-        
-        # Reshape so we can sum every 10 rows
-        mat_reshaped = mat.reshape(n_groups, group_size, mat.shape[1])
-        
-        # Sum across the group_size dimension => shape (n_groups, n_genes)
-        pseudobulk_counts = mat_reshaped.sum(axis=1)
-        
-        # Create an obs DataFrame for these pseudobulks
-        group_ids = [f"{cell_type}_group{i+1}" for i in range(n_groups)]
-        obs_tmp = pd.DataFrame({
-            "cell_types": cell_type,
-            "group_id": group_ids
-        }, index=group_ids)  # index must match the new row names
-        
-        # Store these counts to combine later
-        all_pseudobulks.append(pseudobulk_counts)
-        obs_list.append(obs_tmp)
-    
-    # If nothing was aggregated, return empty
-    if not all_pseudobulks:
-        print("No groups formed for any cell type.")
-        return None
-    
-    # Concatenate everything
-    big_matrix = np.vstack(all_pseudobulks)
-    big_obs = pd.concat(obs_list, axis=0)
-    big_obs.index.name = None  # Clean up
-    
-    # Create new AnnData
-    pseudo_adata = AnnData(X=csr_matrix(big_matrix),  # keep it sparse to save memory
-                           obs=big_obs,
-                           var=pd.DataFrame(index=var_names))
-    
-    return pseudo_adata
-
 def save_cibersort(pseudobulks_df=[], 
                         proportions_df=[], 
                         adata_sc_ref=[], 
@@ -1250,6 +704,21 @@ def save_bayesprism_pseudobulks(pseudobulks_df, proportions_df, output_path):
     proportions_df.index = sample_ids  # Match proportions index to sample IDs
     proportions_df.to_csv(os.path.join(output_path, "proportions.csv"))
 
+def save_bayesprism_realbulks(bulks_df, output_path):
+    """
+    Save pseudobulks for BayesPrism format:
+    - Genes in rows
+    - Samples in columns (indexed from 1 to n)
+    """
+    # Ensure genes are in rows and samples in columns
+    bulks_df = bulks_df.T  # Transpose to get genes in rows, samples in columns
+    print("Bulks shape: ", bulks_df.shape)
+
+    # Assign sequential sample indices starting from 1
+    sample_ids = range(1, bulks_df.shape[1] + 1)
+    bulks_df.columns = sample_ids  # Columns should now be sample IDs
+    bulks_df.to_csv(os.path.join(output_path, "processed_bulks.csv"))
+
 def save_comparison_pseudobulks(pseudobulks_df, proportions_df, output_path, name):
     """
     Save pseudobulks for later comparison with real bulks
@@ -1294,214 +763,6 @@ def save_bayesprism_references(adata, output_path, prefix):
         "tumor_state": 0  # All cells are non-tumor (0)
     })
     cell_state_df.to_csv(os.path.join(output_path, f"{prefix}_cell_state.csv"), index=False)
-
-def remove_diff_genes(sc_adata, sn_adata, diff_genes):
-    """
-    Removes differentially expressed genes from the AnnData objects.
-    If no differentially expressed genes are found, return original data.
-    """
-    if not diff_genes:  # Handle the case where no genes are found
-        print("No differentially expressed genes found. Skipping gene removal step.")
-        return sc_adata, sn_adata
-
-    diff_gene_set = list(set(np.concatenate([df.index.values for df in diff_genes.values()])))
-
-    print("This is diff_gene_set completed:", diff_gene_set)
-
-    print("These are the anndatas before filtering:")
-    print(sc_adata.shape)
-    print(sn_adata.shape)
-
-    # Filter the AnnData objects
-    sn_adata_filtered = sn_adata[:, ~np.isin(sn_adata.var_names, diff_gene_set)].copy()
-    sc_adata_filtered = sc_adata[:, ~np.isin(sc_adata.var_names, diff_gene_set)].copy()
-
-    print("These are the anndatas after filtering:")
-    print(sc_adata_filtered.shape)
-    print(sn_adata_filtered.shape)
-
-    return sc_adata_filtered, sn_adata_filtered
-
-def differential_expression_analysis_parallel(sn_adata, sc_adata, num_threads=4, n_cpus_per_thread=8, deseq_alpha=0.001):
-    from multiprocessing import Pool
-
-    common_cell_types = list(set(sn_adata.obs['cell_types']).intersection(sc_adata.obs['cell_types']))
-    print(f"Running DESeq2 in parallel for {len(common_cell_types)} cell types...")
-
-    tasks = []
-    for ct in common_cell_types:
-        # Subset once in the main process
-        sn_subset = sn_adata[sn_adata.obs['cell_types'] == ct].copy()
-        sc_subset = sc_adata[sc_adata.obs['cell_types'] == ct].copy()
-        tasks.append((ct, sn_subset, sc_subset, n_cpus_per_thread, deseq_alpha))
-
-    with Pool(num_threads) as p:
-        results = p.starmap(run_deseq2_for_cell_type, tasks)
-
-    diff_genes = {cell_type: res for cell_type, res in results if res is not None}
-
-    return diff_genes
-
-def differential_expression_analysis(sn_adata, sc_adata, deseq_alpha=0.001):
-    """
-    Perform differential gene expression analysis between corresponding cell types
-    in single-nucleus (sn_adata) and single-cell (sc_adata) AnnData objects.
-
-    Parameters:
-    - sn_adata: AnnData object for single-nucleus data.
-    - sc_adata: AnnData object for single-cell data.
-
-    Returns:
-    - diff_genes: Dictionary with cell types as keys and DataFrames of differentially
-                  expressed genes as values.
-    """
-
-    # Ensure both AnnData objects have the same set of genes
-    common_genes = sn_adata.var_names.intersection(sc_adata.var_names)
-    sn_adata = sn_adata[:, common_genes]
-    sc_adata = sc_adata[:, common_genes]
-
-    # Initialize the dictionary to store results
-    diff_genes = {}
-
-    # Identify common cell types
-    common_cell_types = set(sn_adata.obs['cell_types']).intersection(sc_adata.obs['cell_types'])
-
-    for cell_type in common_cell_types:
-
-        # Subset data for the current cell type
-        sn_cells = sn_adata[sn_adata.obs['cell_types'] == cell_type]
-        sc_cells = sc_adata[sc_adata.obs['cell_types'] == cell_type]
-        print(f"single nucleus cells of cell type {cell_type}")
-        print(sn_cells.shape)
-        print(f"single cell cells of cell type {cell_type}")
-        print(sc_cells.shape)
-
-        # Convert sparse matrix to dense if necessary
-        sn_X = sn_cells.X if issparse(sn_cells.X) else sn_cells.X.tocsc()
-        sc_X = sc_cells.X if issparse(sc_cells.X) else sc_cells.X.tocsc()
-
-        sn_df = pd.DataFrame(sn_X, columns=common_genes)
-        sc_df = pd.DataFrame(sc_X, columns=common_genes)
-
-        # Combine data into a single DataFrame
-        combined_counts = pd.concat([sn_df, sc_df])
-
-        # Create metadata DataFrame
-        metadata = pd.DataFrame({
-            'condition': ['sn'] * sn_cells.n_obs + ['sc'] * sc_cells.n_obs
-        }, index=combined_counts.index)
-
-        # Initialize and run DESeq2 analysis
-        inference = DefaultInference(n_cpus=8)
-        dds = DeseqDataSet(
-            counts=combined_counts.astype(int),
-            metadata=metadata,
-            design_factors='condition', inference=inference
-        )
-        dds.deseq2()
-
-        # Extract results
-        deseq_stats = DeseqStats(
-            dds,
-            contrast=["condition", "sn", "sc"],
-            alpha=deseq_alpha
-        )
-        deseq_stats.summary()
-        results_df = deseq_stats.results_df
-
-        # Filter for significantly differentially expressed genes
-        sig_genes = results_df[(results_df['padj'] < deseq_alpha) & (results_df['log2FoldChange'].abs() > 1)]
-
-        # Store results in the dictionary
-        diff_genes[cell_type] = sig_genes
-
-    return diff_genes
-
-def run_deseq2_for_cell_type(cell_type, sn_adata, sc_adata, n_cpus=8, deseq_alpha=0.001):
-    """Runs DESeq2 for a single cell type without expression-based filtering."""
-    print(f"Running DESeq2 for {cell_type}...")
-
-    # Subset the data for the cell type
-    sn_cells = sn_adata[sn_adata.obs['cell_types'] == cell_type]
-    sc_cells = sc_adata[sc_adata.obs['cell_types'] == cell_type]
-
-    print(f"For {cell_type}: sn_cells={sn_cells.shape}, sc_cells={sc_cells.shape}")
-
-    # Skip if too few cells
-    if sn_cells.shape[0] < 10 or sc_cells.shape[0] < 10:
-        print(f"Skipping {cell_type}: Too few cells for DE analysis. sn_cells={sn_cells.shape[0]}, sc_cells={sc_cells.shape[0]}")
-        return cell_type, None
-
-    # Ensure both datasets have the same set of genes
-    common_genes = sn_adata.var_names.intersection(sc_adata.var_names)
-    print(f"Cell type {cell_type}: {len(common_genes)} common genes before filtering.")
-
-    if len(common_genes) == 0:
-        print(f"Skipping {cell_type}: No common genes found after intersection.")
-        return cell_type, None
-
-    sn_cells = sn_cells[:, common_genes]
-    sc_cells = sc_cells[:, common_genes]
-
-    # Convert sparse matrices to dense **only if necessary**
-    sn_X = sn_cells.X.toarray() if issparse(sn_cells.X) else sn_cells.X
-    sc_X = sc_cells.X.toarray() if issparse(sc_cells.X) else sc_cells.X
-
-    # Ensure matrices are 2D
-    if sn_X.ndim == 1:
-        sn_X = sn_X.reshape(-1, len(common_genes))
-    if sc_X.ndim == 1:
-        sc_X = sc_X.reshape(-1, len(common_genes))
-
-    print(f"After conversion: sn_X shape = {sn_X.shape}, sc_X shape = {sc_X.shape}")
-
-    # Convert to DataFrames
-    try:
-        sn_df = pd.DataFrame(sn_X, index=sn_cells.obs_names, columns=common_genes)
-        sc_df = pd.DataFrame(sc_X, index=sc_cells.obs_names, columns=common_genes)
-    except ValueError as e:
-        print(f"Error creating DataFrame for {cell_type}: {e}")
-        print(f"sn_X shape: {sn_X.shape}, expected columns: {len(common_genes)}")
-        return cell_type, None
-
-    # Combine data into a single DataFrame
-    combined_counts = pd.concat([sn_df, sc_df])
-
-    # Metadata for DESeq2
-    metadata = pd.DataFrame({'condition': ['sn'] * sn_cells.n_obs + ['sc'] * sc_cells.n_obs}, index=combined_counts.index)
-
-    # Drop zero-variance genes (Retaining all others)
-    combined_counts = combined_counts.loc[:, combined_counts.var(axis=0) > 0]
-    print(f"Cell type {cell_type}: {combined_counts.shape[1]} genes after zero-variance filtering.")
-
-    if combined_counts.shape[1] == 0:
-        print(f"Skipping {cell_type}: No valid genes after zero-variance filtering.")
-        return cell_type, None
-
-    print(f"Running DESeq2 for cell type: {cell_type} with {combined_counts.shape[1]} genes.")
-
-    # Run DESeq2
-    try:
-        inference = DefaultInference(n_cpus=n_cpus)
-        dds = DeseqDataSet(counts=combined_counts.astype(int), metadata=metadata, design_factors="condition", inference=inference)
-        dds.deseq2()
-        deseq_stats = DeseqStats(
-            dds,
-            contrast=["condition", "sn", "sc"],  # or ["condition", "sc", "sn"]
-            alpha=deseq_alpha
-        )
-        deseq_stats.summary()
-        results_df = deseq_stats.results_df
-
-        # Apply differential expression significance filtering
-        sig_genes = results_df[(results_df['padj'] < deseq_alpha) & (results_df['log2FoldChange'].abs() > 1)]
-
-        return cell_type, sig_genes
-
-    except Exception as e:
-        print(f"Error running DESeq2 for {cell_type}: {str(e)}")
-        return cell_type, None
 
 def make_pseudobulks(adata, pseudobulk_config, num_cells, noise, cell_types):
 
@@ -1628,7 +889,7 @@ def make_references(adata_sc_ref, adata_sn, max_cells_per_type=None, seed=42, ce
 
     return adata_sc_ref, adata_sn_ref
 
-def pick_cells(sc_adata_ref, sn_adata, sc_adata_pseudos, min_cells_per_type=75):
+def pick_cells(sc_adata_ref, sn_adata, sc_adata_pseudos, min_cells_per_type=50, check_min_cells=True):
     """
     Selects cell types that are present in all three datasets (sc_adata_ref, sn_adata, sc_adata_pseudos)
     and have at least `min_cells_per_type` occurrences in each.
@@ -1673,7 +934,8 @@ def pick_cells(sc_adata_ref, sn_adata, sc_adata_pseudos, min_cells_per_type=75):
     print(sc_ref_counts[common_cell_types])
     print(sc_pseudos_counts[common_cell_types])
     print(sn_counts[common_cell_types])
-    assert len(valid_cell_types) > 3, "Less than 4 cell types found"
+    if check_min_cells:
+        assert len(valid_cell_types) > 3, "Less than 4 cell types found"
 
     return valid_cell_types
 
@@ -1774,7 +1036,7 @@ def pick_best_datasets(sc_adatas, sn_adatas, min_cells_per_type=50, min_cell_typ
 
     return selected_sc, selected_sn
 
-def downsample_cells_by_type(adata, max_cells=1000, cell_type_key="cell_types"):
+def downsample_cells_by_type(adata, max_cells=1500, cell_type_key="cell_types"):
     """
     Downsamples an AnnData object so that each unique cell type (as given by
     the cell_type_key column in adata.obs) has at most max_cells cells.
@@ -1811,25 +1073,6 @@ def downsample_cells_by_type(adata, max_cells=1000, cell_type_key="cell_types"):
     # Return a new AnnData object with the selected cells
     return adata[selected_indices].copy()
 
-def calculate_median_library_size(adata): 
-    ''' Returns median library size of given AnnData'''
-    
-    # Summation across genes => returns total counts per cell
-    totals = adata.X.sum(axis=1)  # shape: (n_cells,) or maybe a sparse (n_cells,1)
-
-    # 1) Convert to a dense 1D numpy array
-    if issparse(totals):
-        # For a (n_cells,1) sparse matrix, do:
-        totals = totals.A.flatten()  
-    else:
-        # Sometimes sum returns a numpy array with shape (n_cells,1)
-        totals = np.array(totals).flatten()  # ensure 1D
-
-    # 2) Now the median is a single float
-    median_lib = np.median(totals)
-    print("Median library size:", median_lib, type(median_lib))
-    return median_lib
-
 def compute_metrics(y_true, y_pred):
     """Compute Pearson correlation and RMSE between two arrays."""
     rmse = root_mean_squared_error(y_true, y_pred)
@@ -1843,30 +1086,60 @@ def plot_controls(string_title="RMSE", string_column="RMSE", string_ylabel="RMSE
             df_controls={}, 
             control_colors={}, 
             dataset_palette={}):
-    # Create the boxplot for controls using our control colors.
-    plt.figure(figsize=(15, 10))
-    sns.boxplot(x='DisplayTransform', y=string_column, data=df_controls, palette=control_colors)
 
-    plt.title(f"Deconvolution Performance with S-Nuc vs. S-Cell Ref.: \n{string_title}", fontsize=46)
-    plt.xlabel("Control", fontsize=44)
-    plt.ylabel(string_ylabel, fontsize=44)
-    plt.xticks(fontsize=30)
-    plt.yticks(fontsize=30)
+    # Create the boxplot for controls using our control colors.
+    plt.figure(figsize=(18, 12))
+    sns.boxplot(x='DisplayTransform', y=string_column, data=df_controls, 
+                        palette=control_colors, 
+                        linewidth=7, 
+                        fliersize=10)
+    plt.title(f"Deconvolution Performance:\n{string_title}", fontsize=56, fontweight="bold")
+    plt.xlabel("Reference Data Type", fontsize=50, fontweight="bold")
+    plt.ylabel(string_ylabel, fontsize=48, fontweight="bold")
+    plt.xticks([0, 1], ["scRNA Ref.", "snRNA Ref."], fontsize=55)
+    if string_column == "RMSE_All":
+        plt.yticks([0, 0.1, 0.2, 0.3, 0.4], fontsize=45)
+    else:
+        plt.yticks([-1, -0.5, 0,  0.5, 1], fontsize=45)
     plt.tight_layout()
     plt.show()
 
     # Create the boxplot for controls using our control colors.
-    plt.figure(figsize=(15, 10))
-    # # Overlay a swarmplot with points colored by dataset.
-    sns.stripplot(x='DisplayTransform', y=string_column, data=df_controls,
-                hue='Dataset', palette=dataset_palette, dodge=True, size=8)
+    fig, ax = plt.subplots(figsize=(18, 12))    
+    sns.boxplot(
+        x='DisplayTransform',
+        y=string_column,
+        hue='Dataset',
+        data=df_controls,
+        palette=dataset_palette,      # dict mapping each dataset name → color
+        linewidth=7,
+        fliersize=10,
+        ax=ax,
+        dodge=True                    # separate positions for each hue
+    )
 
-    plt.title(f"Deconvolution Performance with S-Nuc vs. S-Cell Ref.: \n{string_title}", fontsize=46)
-    plt.xlabel("Control", fontsize=44)
-    plt.xticks(fontsize=30)
-    plt.yticks(fontsize=30)
-    plt.ylabel(string_ylabel, fontsize=44)
-    plt.legend(title='Dataset', title_fontsize=42, bbox_to_anchor=(1.05, 1), loc='upper left', fontsize=40, markerscale=3)
+    handles, labels = ax.get_legend_handles_labels()
+    ax.legend(
+        handles=handles,
+        labels=labels,
+        title='Dataset',
+        title_fontsize=36,            # larger title
+        fontsize=32,                  # larger text
+        loc='upper left',
+        bbox_to_anchor=(1.02, 1)      # move it outside the main plot
+    )
+
+    plt.title(f"Deconvolution Performance:\n{string_title}",
+              fontsize=56, fontweight="bold")
+    plt.xlabel("Reference Data Type",
+               fontsize=50, fontweight="bold")
+    plt.ylabel(string_ylabel,
+               fontsize=48, fontweight="bold")
+    if string_column == "RMSE_All":
+        plt.yticks([0, 0.1, 0.2, 0.3, 0.4], fontsize=45)
+    else:
+        plt.yticks([-1, -0.5, 0,  0.5, 1], fontsize=45)
+    plt.xticks([0, 1], ["scRNA Ref.", "snRNA Ref."], fontsize=55)
     plt.tight_layout()
     plt.show()
 
@@ -1953,3 +1226,60 @@ def scvi_train_model(model_save_path, sc_adata, sn_adata, heldout_label=None, co
         model.save(model_save_path, overwrite=True)
 
     return model
+
+def open_adipose_datasets_all(res_name="Real_ADP", base_dir=".."):
+
+    ''' Open datasets and return them processed and in order ''' 
+
+    # Define file paths
+    bulk_file = f'{base_dir}/data/{res_name}/adipose_bulks_processed.csv'
+    sn_sc_file = f'{base_dir}/data/{res_name}/sc_sn_{res_name}_allcells.h5ad'
+
+    # 1. Read in the data
+    bulk_df = pd.read_csv(bulk_file, index_col=0)
+    print("Bulk data shape:", bulk_df.shape)
+
+    adata = sc.read_h5ad(sn_sc_file)
+    print("Combined SN+SC data shape:", adata.shape)
+
+    # Match genes between datasets
+    # Assume that bulk_df columns are gene names. Identify the common gene set.
+    common_genes = list(set(bulk_df.columns).intersection(set(adata.var_names)))
+    print("Number of common genes:", len(common_genes))
+
+    adata = adata[:, common_genes].copy()
+    bulk_df = bulk_df[common_genes].copy()
+
+    # 3. Separate SN and SC data based on the 'data_type' column
+    sc_adata = adata[adata.obs['data_type'] == 'single_cell'].copy()
+    sn_adata = adata[adata.obs['data_type'] == 'single_nucleus'].copy()
+
+    del adata
+
+    print("Single-cell data shape:", sc_adata.shape)
+    print("Single-nucleus data shape:", sn_adata.shape)
+    print("Bulk data shape:", bulk_df.shape)
+    print("Cell Type counts:")
+    print(sc_adata.obs.cell_types.value_counts())
+    print(sn_adata.obs.cell_types.value_counts())
+
+    matching = np.intersect1d(sn_adata.obs.cell_types.unique(), sc_adata.obs.cell_types.unique())
+    matching
+
+    nonmatching = np.setxor1d(sn_adata.obs.cell_types.unique(), sc_adata.obs.cell_types.unique())
+    nonmatching
+
+    sn_adata_match = sn_adata[sn_adata.obs.cell_types.isin(matching)].copy()
+
+    sn_missing={}
+    sn_miss_adata = sn_adata[sn_adata.obs.cell_types.isin(nonmatching)].copy()
+    sn_missing[0] = sn_miss_adata[sn_miss_adata.obs.cell_types == nonmatching[0]].copy()
+    sn_missing[1] = sn_miss_adata[sn_miss_adata.obs.cell_types == nonmatching[1]].copy()
+
+    print(sn_adata_match.obs.cell_types.value_counts())
+
+    sn_adata = sn_adata_match.copy()
+
+    print(sn_miss_adata.obs.cell_types.value_counts())
+
+    return bulk_df, sc_adata, sn_adata, sn_missing
